refactor(api): extract route registration from NewRouter (#565 [39], part 2)
Completes finding [39]. NewRouter mixed DI, job orchestration, config
reload, and ~780 lines of route registration (~30 nested gin route
groups) in one function. Findings [40] and [39]-part-1 already
addressed the first two; this extracts the remaining route table into
two new functions in router_routes.go:

- registerPublicRoutes: the unauthenticated Terraform-protocol/OCI/
  Swagger surface (/v2, /v1/modules, /v1/providers, /terraform/*).
- registerAPIV1Routes: the /api/v1, /scim/v2, /dev, and webhook route
  table.

Both function bodies are the ORIGINAL route-registration code moved
verbatim -- mechanically verified via a whitespace-normalized substring
match against the pre-extraction text, not rewritten. Each takes a
small dependency struct (publicRouteDeps: 12 fields, apiV1RouteDeps: 52
fields) built from NewRouter's already-constructed locals; every field
is re-bound to a local of the identical name at the top of the function
(e.g. `cfg := d.cfg`) so the moved body needed zero further edits beyond
that header.

NewRouter shrinks from ~1200 lines to ~490 -- now DI construction, job
registration, and two route-table calls, matching the god-function
complexity concern (CWE-1120) the finding raised.

Verified before and after the extraction:
- Every field in both deps structs is set in NewRouter's call-site
  literal (cross-checked as an exact set match, 12/12 and 52/52) -- a
  missing field wouldn't be a compile error, only a silent nil/zero at
  runtime, so this needed explicit checking beyond `go build`.
- No variable name backing a struct field is declared more than once
  anywhere in router.go (ruling out shadowing silently substituting the
  wrong value for a same-named field).
- registerPublicRoutes/registerAPIV1Routes are called at the exact same
  point in NewRouter's sequence the original inline code ran (verified
  against the pre-extraction diff) -- no reordering relative to
  jobRegistry.StartAll, egressGuard setup, or anything else that stayed
  in NewRouter.
- go build/vet/gofmt/golangci-lint (0 issues)/gosec (0 new findings)/
  full test suite all clean.
